Update price type

+++ migrations/001_init.up.sql
@@ -11,7 +11,7 @@ CREATE TABLE IF NOT EXISTS ads (
     title VARCHAR(100) NOT NULL,
     description TEXT NOT NULL,
     image_url TEXT NOT NULL,
-    price DECIMAL(10, 2) NOT NULL,
+    price INTEGER NOT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMP NOT NULL DEFAULT NOW(),
     user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE
